alpha of cosmos refactor... we hope
